feat(ssa4yak literal) refacoter literal stmt and support templatestring
